+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Filter watts for crazy values. Only >0 && <3000 now accepted. 
 - Fixed bug where scans may not happen even when configured devices aren't connected.  
 - Worked with Mark Roy to tune PID.
+- Added proper rounding from float to int for power and cadence.
+- More ERG tweaks for Marc Roy. 
+- If homed, we throw out negative PowerTable returns. 
 
 ### Hardware